requirements.txt
Django==1.8.3
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-annoying==0.8.3
geopy==1.10.0
gunicorn==19.3.0
httplib2==0.9.1
ipython==3.2.1
python-instagram==1.3.2
simplejson==3.8.0
six==1.9.0
static3==0.6.1
wheel==0.24.0
wsgiref==0.1.2
